vimrc: updated to handle using tabs better
No longer using YouCompleteMe, supertab
is simpler and the functionality of YCM
is kind of annoying with compile errors.

Also move Ctrl-P working path to be the top
directory with a .git/.svn.

Added shortcut mappings for navigating tabs,
but the one for :buffers doesn't seem to work.
Maybe another key is also binded to that already?