tmux.conf: added zoom functionality for copy-pasting
tmux messes with terminal's normal copy and paste features
because terminal cannot recognize tmux | symbols as the borders.
I read some blog and this was their somewhat hacky way around that,
which moves the pane in question to a new pane, and then you can
perform normal copy-paste functions there, and then zoom back
to the pane. Hacky, but I think it works surprisingly well.