ctrlp optimize custom ignore
Added more ignores for the custom ignore. I read online that if you use
a ctrlp_user_command instead of the default globpath() so therefore you
need to unlet the ctrlp_user_command if you want your filters to
actually work. However, manage does a good job following generic repos
in which the .gitignore ignores most of the bulk files, so we should not
remove the ctrlp_user_command from running be default. Only the router
checkout needs to be explicitly ignored, since it's poorly managed in
terms of the extra things that are in it.

Also added some new conditions to the router checkout filters since
there are so many things in the router checkout. Seemed to have improved
things quite a bit.