add zshrc for mac and linux
Started using zsh instead of bash. It seems to be slightly nicer, but
the cross compatibility is a bit annoying. Most of the magic is in
oh-my-zsh, which you can download here.

https://github.com/robbyrussell/oh-my-zsh

If you are using iterm2, you need a powerline font for non-ascii
characters. The one that I found that works can be downloaded here:

https://github.com/powerline/fonts/blob/master/Meslo/Meslo%20LG%20M%20DZ%20Regular%20for%20Powerline.otf

Make this 11pt font just like the text, and you can still use the normal
font you like.

Note: I commit two different zshrc* files but they are almost exactly
same, just that the directory pathing is slightly different. Oh wells.