tmux.conf: updated default window opening path when making new windows
Noticed this when switching from tmux1.8 to tmux1.9, but now by default,
window/panes are not opened in the previous pane's current path. This is
really annoying if you are at a path that you don't want to type again,
so this fixes the issue.

Also, while learning more about tmux, you can run 'tmux list-keys' or
use the command seq + '?' to see all the key commands, which is a pretty
useful list to consult sometimes.